pyproject.toml
[tool.poetry]
package-mode = false
name = "flask-live-transcription"
version = "0.1.0"
description = ""
authors = ["Jacob-Lasky <jacob.s.lasky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
deepgram-sdk = "v3.7.5"
flask = "3.0.0"
flask-socketio = "5.3.6"
python-dotenv = "1.0.0"
requests = "^2.32.3"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
